backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-groq
pinecone-client
pydantic
